Semi-completed player combat
The player is able to attack with one dedicated animation and send information to the hostile if detected. It is very rigid however, so the chances of multiple weapons is completely off the table. We should be able to add some animation counters and perhaps an aerial animation to the player soon